requirements.txt
Django==5.2.7
djangorestframework==3.16.1
gunicorn==23.0.0
pandas==2.3.3
numpy==2.3.4
matplotlib==3.10.7
liac-arff==2.5.0
seaborn==0.13.2
sqlparse==0.5.3
asgiref==3.10.0
pytz==2025.2
tzdata==2025.2
